Improve node packing
Move the size from the children vector to the node to avoid unfortunate
padding, and add view types to still present a standard vector interface.

+++ include/spaix/sizes.hpp
@@ -44,7 +44,7 @@ power(const T b, const T e) noexcept
    elements, respectively.
 */
 template<class DirectoryNodeType, class DataNodeType, DataPlacement placement>
-constexpr size_t
+constexpr unsigned
 max_tree_height(const ChildCount min_fanout) noexcept
 {
   constexpr auto total_space = std::numeric_limits<size_t>::max();
@@ -53,7 +53,7 @@ max_tree_height(const ChildCount min_fanout) noexcept
 
   if constexpr (placement == DataPlacement::inlined) {
     const auto n_most_dirs = total_space / dir_size;
-    return log_b(n_most_dirs, size_t{min_fanout});
+    return static_cast<unsigned>(log_b(n_most_dirs, size_t{min_fanout}));
   }
 
   const auto n_most_dats         = total_space / dat_size;
@@ -62,7 +62,7 @@ max_tree_height(const ChildCount min_fanout) noexcept
   const auto available_dat_space = total_space - needed_dir_space;
   const auto max_n_dats          = available_dat_space / dat_size;
 
-  return log_b(max_n_dats, size_t{min_fanout});
+  return static_cast<unsigned>(log_b(max_n_dats, size_t{min_fanout}));
 }
 
 } // namespace spaix
